test(client): centralize setupAndGetUserPage, known searches and users
